Clean up Domains CMR stuff
