+++ packages/manager/src/features/Domains/DomainsLanding.tsx
@@ -7,11 +7,8 @@ import { connect, MapStateToProps } from 'react-redux';
 import { RouteComponentProps } from 'react-router-dom';
 import { compose } from 'recompose';
 import DomainIcon from 'src/assets/icons/entityIcons/domain.svg';
-import AddNewLink from 'src/components/AddNewLink';
-import Breadcrumb from 'src/components/Breadcrumb';
 import Button from 'src/components/Button';
 import CircleProgress from 'src/components/CircleProgress';
-import FormControlLabel from 'src/components/core/FormControlLabel';
 import {
   createStyles,
   Theme,
@@ -25,20 +22,15 @@ import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { EntityTableRow, HeaderCell } from 'src/components/EntityTable';
 import EntityTable_CMR from 'src/components/EntityTable/EntityTable_CMR';
 import ErrorState from 'src/components/ErrorState';
-import Grid from 'src/components/Grid';
 import LandingHeader from 'src/components/LandingHeader';
 import Link from 'src/components/Link';
 import Notice from 'src/components/Notice';
 import { Order } from 'src/components/Pagey';
 import Placeholder from 'src/components/Placeholder';
 import PreferenceToggle, { ToggleProps } from 'src/components/PreferenceToggle';
-import Toggle from 'src/components/Toggle';
 import domainsContainer, {
   Props as DomainProps
 } from 'src/containers/domains.container';
-import withFeatureFlags, {
-  FeatureFlagConsumerProps
-} from 'src/containers/withFeatureFlagConsumer.container.ts';
 import { Domains } from 'src/documentation';
 import { ApplicationState } from 'src/store';
 import {
@@ -53,7 +45,6 @@ import { sendGroupByTagEnabledEvent } from 'src/utilities/ga';
 import DisableDomainDialog from './DisableDomainDialog';
 import { Handlers as DomainHandlers } from './DomainActionMenu';
 import DomainRow from './DomainTableRow';
-import DomainRow_CMR from './DomainTableRow_CMR';
 import DomainZoneImportDrawer from './DomainZoneImportDrawer';
 
 const DOMAIN_CREATE_ROUTE = '/domains/create';
@@ -67,7 +58,8 @@ type ClassNames =
   | 'dnsWarning'
   | 'tagWrapper'
   | 'tagGroup'
-  | 'importButton';
+  | 'importButton'
+  | 'banner';
 
 const styles = (theme: Theme) =>
   createStyles({
@@ -101,6 +93,9 @@ const styles = (theme: Theme) =>
     importButton: {
       marginLeft: -theme.spacing(),
       whiteSpace: 'nowrap'
+    },
+    banner: {
+      marginBottom: theme.spacing()
     }
   });
 
@@ -138,8 +133,7 @@ export type CombinedProps = DomainProps &
   RouteComponentProps<{}> &
   StateProps &
   DispatchProps &
-  WithSnackbarProps &
-  FeatureFlagConsumerProps;
+  WithSnackbarProps;
 
 const headers: HeaderCell[] = [
   {
@@ -328,7 +322,6 @@ export class DomainsLanding extends React.Component<CombinedProps, State> {
       domainsData,
       domainsLoading,
       domainsLastUpdated,
-      flags,
       howManyLinodesOnAccount,
       isLargeAccount,
       isRestrictedUser,
@@ -343,7 +336,7 @@ export class DomainsLanding extends React.Component<CombinedProps, State> {
     };
 
     const domainRow: EntityTableRow<Domain> = {
-      Component: flags.cmr ? DomainRow_CMR : DomainRow,
+      Component: DomainRow,
       data: domainsData ?? [],
       request: isLargeAccount ? getDomains : undefined,
       handlers,
@@ -411,14 +404,12 @@ export class DomainsLanding extends React.Component<CombinedProps, State> {
         <DocumentTitleSegment segment="Domains" />
         {shouldShowBanner && (
           <Notice warning important className={classes.dnsWarning}>
-            <Typography variant="h3">
-              Your DNS zones are not being served.
-            </Typography>
-            <Typography>
-              Your domains will not be served by Linode&#39;s nameservers unless
-              you have at least one active Linode on your account.
-              <Link to="/linodes/create"> You can create one here.</Link>
-            </Typography>
+            <div className={classes.banner}>
+              <strong>Your DNS zones are not being served.</strong>
+            </div>
+            Your domains will not be served by Linode&#39;s nameservers unless
+            you have at least one active Linode on your account.
+            <Link to="/linodes/create"> You can create one here.</Link>
           </Notice>
         )}
         {this.props.location.state?.recordError && (
@@ -438,79 +429,21 @@ export class DomainsLanding extends React.Component<CombinedProps, State> {
           }: ToggleProps<boolean>) => {
             return (
               <React.Fragment>
-                {flags.cmr ? (
-                  <LandingHeader
-                    title="Domains"
-                    extraActions={
-                      <Button
-                        className={classes.importButton}
-                        onClick={this.openImportZoneDrawer}
-                        buttonType="secondary"
-                      >
-                        Import a Zone
-                      </Button>
-                    }
-                    entity="Domain"
-                    onAddNew={this.navigateToCreate}
-                    docsLink="https://www.linode.com/docs/platform/manager/dns-manager/"
-                  />
-                ) : (
-                  <Grid
-                    container
-                    justify="space-between"
-                    alignItems="flex-end"
-                    style={{ paddingBottom: 0 }}
-                  >
-                    <Grid item className={classes.titleWrapper}>
-                      <Breadcrumb
-                        // This component can be rendered with the URL
-                        // /domains/:domainId, which would result in a double
-                        // breadcrumb. Thus we give the <Breadcrumb /> an explicit
-                        // pathname.
-                        pathname="Domains"
-                        labelTitle="Domains"
-                        className={classes.breadcrumbs}
-                      />
-                    </Grid>
-                    <Grid item className="p0">
-                      <FormControlLabel
-                        className={classes.tagGroup}
-                        control={
-                          <Toggle
-                            className={
-                              domainsAreGrouped ? ' checked' : ' unchecked'
-                            }
-                            onChange={toggleGroupDomains}
-                            checked={domainsAreGrouped}
-                            disabled={isLargeAccount}
-                          />
-                        }
-                        label="Group by Tag:"
-                      />
-                    </Grid>
-                    <Grid item>
-                      <Grid
-                        container
-                        alignItems="flex-end"
-                        // style={{ width: 'auto' }}
-                      >
-                        <Grid item className="pt0">
-                          <AddNewLink
-                            onClick={this.openImportZoneDrawer}
-                            label="Import a Zone"
-                          />
-                        </Grid>
-                        <Grid item className="pt0">
-                          <AddNewLink
-                            data-testid="create-domain"
-                            onClick={this.navigateToCreate}
-                            label="Add a Domain"
-                          />
-                        </Grid>
-                      </Grid>
-                    </Grid>
-                  </Grid>
-                )}
+                <LandingHeader
+                  title="Domains"
+                  extraActions={
+                    <Button
+                      className={classes.importButton}
+                      onClick={this.openImportZoneDrawer}
+                      buttonType="secondary"
+                    >
+                      Import a Zone
+                    </Button>
+                  }
+                  entity="Domain"
+                  onAddNew={this.navigateToCreate}
+                  docsLink="https://www.linode.com/docs/platform/manager/dns-manager/"
+                />
                 <EntityTable_CMR
                   entity="domain"
                   toggleGroupByTag={toggleGroupDomains}
@@ -660,7 +593,6 @@ export default compose<CombinedProps, Props>(
   domainsContainer(),
   connected,
   withSnackbar,
-  withFeatureFlags,
   styled
 )(DomainsLanding);
